pyproject.toml
[project]
name = "automated-changelog-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "ipykernel>=7.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
